+++ lib/minfraud/model/subscores.rb
@@ -59,6 +59,11 @@ module Minfraud
       # @return [Float, nil]
       attr_reader :cvv_result
 
+      # The risk associated with the device. If present, this is a value in the
+      # range of 0.01 to 99.
+      # @return [Float, nil]
+      attr_reader :device
+
       # The risk associated with the particular email address. If present, this
       # is a value in the range 0.01 to 99.
       #
@@ -70,6 +75,12 @@ module Minfraud
       #
       # @return [Float, nil]
       attr_reader :email_domain
+
+      # The risk associated with the email address local part (the part of
+      # the email address before the @ symbol). If present, this is a value
+      # in the range 0.01 to 99.
+      # @return [Float, nil]
+      attr_reader :email_local_part
 
       # The risk associated with the issuer ID number on the email domain. If
       # present, this is a value in the range 0.01 to 99.
@@ -111,6 +122,11 @@ module Minfraud
       # @return [Float, nil]
       attr_reader :phone_number
 
+      # The risk associated with the shipping address. If present, this is a
+      # value in the range 0.01 to 99.
+      # @return [Float, nil]
+      attr_reader :shipping_address
+
       # The risk associated with the distance between the shipping address and
       # the IP location for the given IP address. If present, this is a value
       # in the range 0.01 to 99.
@@ -139,13 +155,16 @@ module Minfraud
         @country                                  = get('country')
         @country_mismatch                         = get('country_mismatch')
         @cvv_result                               = get('cvv_result')
+        @device                                   = get('device')
         @email_address                            = get('email_address')
         @email_domain                             = get('email_domain')
+        @email_local_part                         = get('email_local_part')
         @email_tenure                             = get('email_tenure')
         @ip_tenure                                = get('ip_tenure')
         @issuer_id_number                         = get('issuer_id_number')
         @order_amount                             = get('order_amount')
         @phone_number                             = get('phone_number')
+        @shipping_address                         = get('shipping_address')
         @shipping_address_distance_to_ip_location = get(
           'shipping_address_distance_to_ip_location'
         )