+++ spec/model/factors_spec.rb
@@ -21,13 +21,16 @@ describe Minfraud::Model::Factors do
       expect(m.subscores.country).to eq 0.06
       expect(m.subscores.country_mismatch).to eq 0.07
       expect(m.subscores.cvv_result).to eq 0.08
+      expect(m.subscores.device).to eq 0.18
       expect(m.subscores.email_address).to eq 0.09
       expect(m.subscores.email_domain).to eq 0.10
+      expect(m.subscores.email_local_part).to eq 0.19
       expect(m.subscores.email_tenure).to eq 0.11
       expect(m.subscores.ip_tenure).to eq 0.12
       expect(m.subscores.issuer_id_number).to eq 0.13
       expect(m.subscores.order_amount).to eq 0.14
       expect(m.subscores.phone_number).to eq 0.15
+      expect(m.subscores.shipping_address).to eq 0.2
       expect(m.subscores.shipping_address_distance_to_ip_location).to eq 0.16
       expect(m.subscores.time_of_day).to eq 0.17
     end